--- modules/pipeline/endpoints/http-client.private.env.json
@@ -1,5 +1,0 @@
-{
-  "erda-dev": {
-    "addr": "localhost:3081"
-  }
-}
